pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-mistralai>=1.1.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
